dev: decouple hyper
